requirements.txt
asgiref==3.5.0
Django==4.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.2
tzdata==2023.3
